Updated SequenceGame and Tile
Implemented functionality for the user to choose between four difficulties or exit. This works properly with restarting the game after having lost.

I also changed the Ellipses back to Rectangles in Tile for debugging purposes.

Hopefully you haven't started working on fixing the issue where the user can click while the sequence is animating and you can just work off of this :)